fix: issue with storybook
