+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "publicsuqare-clone",
+  "name": "avi-proj",
   "version": "0.1.0",
   "private": true,
   "license": "private",
