+++ stories/assets/figma-plugin.png
